Refactor utility functions out of component
