+++ components/listings/shared/ListingFilter/components/FilterButton/styles.js
@@ -1,21 +1,15 @@
 import styled from 'styled-components'
 import theme from '@emcasa/ui'
 import Button from '@emcasa/ui-dom/components/Button'
-import Icon from '@emcasa/ui-dom/components/Icon'
 
 const StyledFilterButton = styled(Button)`
   border: 1px solid ${theme.colors.pink};
   color: ${({active}) => active ? theme.colors.white : theme.colors.pink};
   font-size: ${theme.fontSizes[1]}px;
   margin: 0 ${theme.space[2]}px ${theme.space[1]}px 0;
-`
-
-const RemoveIcon = styled(Icon)`
-  display: flex;
-  margin-left: ${theme.space[2]}px;
+  height: 32px;
 `
 
 export {
-  StyledFilterButton,
-  RemoveIcon
+  StyledFilterButton
 }
