+++ components/listings/shared/ListingFilter/components/FilterPanel/index.js
@@ -22,8 +22,8 @@ class FilterPanel extends PureComponent {
           {this.props.children}
         </Row>
         <Row justifyContent="space-between" mt={2}>
-          <Button link onClick={this.props.clear}>Limpar</Button>
-          <Button link onClick={this.props.apply}>Aplicar</Button>
+          <Button p={0} link style={{color: theme.colors.dark, height: 32}} onClick={this.props.clear}>Limpar</Button>
+          <Button p={0} link style={{height: 32}} onClick={this.props.apply}>Aplicar</Button>
         </Row>
       </Container>
     )
